+++ packages/commander-zod-prompt/jest.config.js
@@ -1,0 +1,7 @@
+const baseConfigFn = require('../../jest.base.config');
+
+/** @type {import('@jest/types').Config.InitialOptions} */
+module.exports = {
+  ...baseConfigFn('commander-zod-prompt'),
+  displayName: 'commander-zod-prompt',
+};