+++ packages/commander-zod/src/lib/types.ts
@@ -41,26 +41,25 @@ export type CommandDefinition = {
    */
   environmentPrefix?: string;
 
+  /** When true, help descriptions will also include the config key for all parameters */
+  includeConfigNameInParameterDescriptions?: boolean;
+
   /** Parameter definitions
    *
-   * Supports both arguments and optional parameters. Despite the name
-   * "optional", both parameter types can be made mandatory. The main
-   * difference is that arguments are positional while options include
-   * a key (i.e. `--key value`)
+   * - The `key` given for the parameter will be used as the variable
+   * name for the resolved value. All resolved parameters will then
+   * be passed as a `props` object into an `action` and/or `hook` callback.
+   * The types will be statically resolved and inferred based on the
+   * zod schema defined. If no zod schema is defined a string is assumed.
    */
   parameters?: {
-    /** The name of the parameter
-     *
-     * If a value is found for this parameter it will be passed
-     * to the action handler
-     */
     [key: string]: ParameterDefinition;
   };
 };
 
 export type ParameterDefinition = ArgumentDefinition | OptionDefinition;
 
-interface BaseDefinition {
+export interface BaseDefinition {
   /** Zod Schema to use for validation and type inference */
   schema?: z.ZodFirstPartySchemaTypes;
 
@@ -80,13 +79,19 @@ interface BaseDefinition {
   /** List of valid choices for the parameter */
   choices?: string[];
 
-  /** Invidual parameter environment flag
+  /** Individual parameter environment flag
    *
-   * This can be a boolean to use the default environment name translation
+   * - When false, this disables the environment variable for this parameter
+   * (true by default)
+   *
+   * - This can be a boolean to use the default environment name translation
    * or a string to use as the environment name. The latter is a convenience
    * over having to also override the parameter in the `names` property
    */
   environment?: boolean | string;
+
+  /** When false, it will disable populating this value from a config (true by default) */
+  useConfig?: boolean;
 
   /** Parameter names for different sources.
    *